+++ lib/pragma/src/config.cpp
@@ -1,5 +1,5 @@
 // Metashell - Interactive C++ template metaprogramming shell
-// Copyright (C) 2017, Abel Sinkovics (abel@sinkovics.hu)
+// Copyright (C) 2019, Abel Sinkovics (abel@sinkovics.hu)
 //
 // This program is free software: you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
@@ -17,30 +17,94 @@
 #include <metashell/pragma/config.hpp>
 
 #include <boost/algorithm/string/join.hpp>
+#include <boost/range/adaptor/filtered.hpp>
 #include <boost/range/adaptor/transformed.hpp>
+#include <boost/range/adaptor/uniqued.hpp>
+
+#include <boost/optional.hpp>
+
+#include <algorithm>
+#include <set>
 
 namespace metashell
 {
   namespace pragma
   {
-    std::string config::description() const
+    namespace
     {
-      return "Lists all available configs.";
+      boost::optional<std::string>
+      element_to_show(const boost::optional<data::shell_config_name>& prefix_,
+                      const data::shell_config_name& name_)
+      {
+        if (!prefix_)
+        {
+          return to_string_with_standard_suffix(name_.front());
+        }
+
+        if (size(name_) <= size(*prefix_))
+        {
+          return boost::none;
+        }
+
+        auto i = name_.begin();
+        for (const auto& e : *prefix_)
+        {
+          if (e.value() != i->value())
+          {
+            return boost::none;
+          }
+          ++i;
+        }
+
+        return to_string_with_standard_suffix(*i);
+      };
     }
 
-    void config::run(iface::shell& shell_, iface::displayer& displayer_) const
+    std::string config::arguments() const { return "[<name>]"; }
+
+    std::string config::description() const
+    {
+      return "Lists the available configs and config folders in config folder "
+             "<name>.";
+    }
+
+    void config::run(const data::command::iterator&,
+                     const data::command::iterator&,
+                     const data::command::iterator& args_begin_,
+                     const data::command::iterator& args_end_,
+                     iface::shell& shell_,
+                     iface::displayer& displayer_) const
     {
       using boost::algorithm::join;
       using boost::adaptors::transformed;
+      using boost::adaptors::filtered;
 
       const data::config& cfg = shell_.get_config();
-      const data::shell_config* active = &cfg.active_shell_config();
+      const data::shell_config_name active = cfg.active_shell_config().name;
+
+      const auto arg = tokens_to_string(args_begin_, args_end_);
+      const boost::optional<data::shell_config_name> prefix =
+          empty(arg) ?
+              boost::none :
+              boost::make_optional(data::shell_config_name(arg.value()));
+
+      const auto active_to_show = element_to_show(prefix, active);
+
+      const auto element_range =
+          cfg.shell_configs() |
+          transformed([&prefix](const data::shell_config& cfg_) {
+            return element_to_show(prefix, cfg_.name);
+          }) |
+          filtered([](const auto& element_) { return bool(element_); }) |
+          transformed([](const auto& element_) { return *element_; });
+
+      const std::set<std::string> elements(
+          element_range.begin(), element_range.end());
 
       displayer_.show_comment(data::text(
-          join(cfg.shell_configs() |
-                   transformed([active](const data::shell_config& cfg_) {
-                     return (&cfg_ == active ? " * " : "   ") + cfg_.name;
-                   }),
+          join(elements | transformed([&active_to_show](const auto& element_) {
+                 return (element_ == active_to_show ? " * " : "   ") + element_;
+               }),
                "\n")));
     }
   }